+++ gatsby-config.js
@@ -81,9 +81,11 @@ module.exports = {
       resolve: 'gatsby-plugin-netlify',
     },
     {
-      resolve: 'gatsby-plugin-google-gtag',
+      resolve: 'gatsby-plugin-google-tagmanager',
       options: {
-        trackingIds: ['G-6EQKK25267'],
+        id: 'GTM-P33RLCP',
+        includeInDevelopment: true,
+        routeChangeEventName: 'gatsby-route-change',
       },
     },
     {