+++ src/constants.js
@@ -7,3 +7,8 @@ export const INITIAL_COLOR_MODE_CSS_PROP = '--initial-color-mode';
 export const AUTHOR_SLUG = 'author';
 export const BLOG_SLUG = 'blog';
 export const CATEGORY_SLUG = 'category';
+
+export const COOKIES_ACCEPTED_COOKIE_NAME = 'visiosto_cookies_accepted';
+export const DISABLE_ANALYTICS_COOKIE_NAME = 'visiosto_analytics_opt_out';
+
+export const GOOGLE_ANALYTICS_TRACKING_ID = 'G-6EQKK25267';