+++ src/locales/en.js
@@ -6,6 +6,7 @@ import { defineMessages } from 'react-intl';
 import authorMessages from './en/author';
 import blogMessages from './en/blog';
 import categoryMessages from './en/category';
+import cookieNoticeMessages from './en/cookie-notice';
 import footerMessages from './en/footer';
 import headerMessages from './en/header';
 import indexMessages from './en/index';
@@ -19,6 +20,7 @@ export const messages = {
   ...authorMessages,
   ...blogMessages,
   ...categoryMessages,
+  ...cookieNoticeMessages,
   ...footerMessages,
   ...headerMessages,
   ...indexMessages,