+++ src/locales/en/cookie-notice.js
@@ -1,0 +1,24 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+export default {
+  cookieSettingsLink: 'Cookie settings',
+  // Fancy a cookie?
+  cookieNoticeTitle: 'Would you like a cookie?',
+  cookieNoticeDescription:
+    'Our website uses cookies to make your experience with us even sweeter. We use cookies to see what content our visitors are interested in and what we could do better. You can learn more about the cookies we use and manage your settings on the cookie settings page.',
+  cookieNoticeAccept: 'Thanks, I accept cookies',
+  cookieNoticeReject: 'I’ll check out the cookies and adjust them',
+  cookieNoticeSettings: 'Settings',
+  cookieNoticeFunctionalTitle: 'Functional cookies',
+  cookieNoticeFunctionalDescription:
+    'These cookies are required for our website to work properly and cannot be turned off. Among other things, they take care of remembering the cookie preferences you select and loading different functions on the website, such as maps and online tests.',
+  cookieNoticeAnonymizedTrackingTitle: 'Anonymized tracking cookies',
+  cookieNoticeGoogleAnalytics: 'Google Analytics',
+  cookieNoticeGoogleAnalyticsInfo: 'What does this do?',
+  cookieNoticeAnonymizedTrackingDescription:
+    'Tracking cookies help us track what content our visitors are interested in and where we need to do better. We cannot identify you with these cookies.',
+  cookieNoticeDataProtection: 'Data protection',
+  cookieNoticeCancel: 'Cancel',
+  cookieNoticeSave: 'Save and close',
+};