+++ src/locales/fi/cookie-notice.js
@@ -1,0 +1,24 @@
+// Copyright (c) 2021 Visiosto oy
+// Licensed under the MIT License
+
+export default {
+  cookieSettingsLink: 'Evästeasetukset',
+  // Fancy a cookie?
+  cookieNoticeTitle: 'Saisiko olla keksi?',
+  cookieNoticeDescription:
+    'Sivustomme käyttää evästeitä. Käytämme evästeitä nähdäksemme mistä sisällöistä sivustomme käyttäjät ovat kiinnostuneita ja mahdollistaaksemme joitakin sivuston toiminnallisuuksia. Voit tutustua tarkemmin evästeiden sisältöön ja hallita asetuksiasi evästeasetus-sivulla.',
+  cookieNoticeAccept: 'Kiitos, hyväksyn evästeet',
+  cookieNoticeReject: 'Tarkistan evästeet ja säädän niitä',
+  cookieNoticeSettings: 'Asetukset',
+  cookieNoticeFunctionalTitle: 'Toiminnalliset evästeet',
+  cookieNoticeFunctionalDescription:
+    'Toiminnallisilla evästeillä mahdollistamme erilaisten sisältöjen ja toiminnallisuuksien (kuten karttapalveluiden ja kyselyiden lisäämisen helposti osaksi sivustoa). Toiminnallisia evästeitä ei voi laittaa pois päältä, sillä niitä tarvitaan sivuston toiminnan mahdollistamiseksi.',
+  cookieNoticeAnonymizedTrackingTitle: 'Anonymisoidut seurantaevästeet',
+  cookieNoticeGoogleAnalytics: 'Google Analytics',
+  cookieNoticeGoogleAnalyticsInfo: 'Mitä tämä tekee?',
+  cookieNoticeAnonymizedTrackingDescription:
+    'Seurantaevästeiden avulla seuraamme mitkä sisällöt kävijöitämme kiinnostavat ja missä meillä on syytä petrata. Emme voi tunnistaa sinua evästeiden avulla.',
+  cookieNoticeDataProtection: 'Tietosuoja',
+  cookieNoticeCancel: 'Peruuta',
+  cookieNoticeSave: 'Tallenna ja sulje',
+};